Update styling docs for selective import and extract issue list
Rewrite the settings doc and styling-file-format ADR import semantics to
describe the selective modal (the confirm-dialog merge flow and getStylingConflicts
are gone). Port the Suspense-fallback regression test into the modal's test file,
and extract the parse-issue rendering into its own ImportIssuesList component.
