+++ packages/graph-explorer/src/modules/StyleImport/ImportIssuesList.tsx
@@ -1,0 +1,100 @@
+import type {
+  EntryImportIssue,
+  GeneralImportIssue,
+  ImportIssue,
+} from "@/core/styling";
+
+import {
+  AlertDialogBody,
+  Group,
+  GroupHeader,
+  GroupItem,
+  GroupTitle,
+} from "@/components";
+
+/**
+ * Renders the per-field issues from a rejected styling file, grouped into file
+ * structure, node types, and edge types. Import is atomic, so this is shown
+ * instead of the modal when any value fails validation.
+ */
+export function ImportIssuesList({ issues }: { issues: ImportIssue[] }) {
+  const general = issues.filter(issue => issue.scope === "general");
+  const entries = issues.filter(issue => issue.scope === "entry");
+  const vertices = entries.filter(issue => issue.entityType === "vertex");
+  const edges = entries.filter(issue => issue.entityType === "edge");
+  return (
+    <AlertDialogBody className="space-y-3">
+      {general.length > 0 ? <GeneralIssuesGroup issues={general} /> : null}
+      {vertices.length > 0 ? (
+        <EntryIssuesGroup
+          label={`${vertices.length} node ${vertices.length === 1 ? "type" : "types"}`}
+          issues={vertices}
+        />
+      ) : null}
+      {edges.length > 0 ? (
+        <EntryIssuesGroup
+          label={`${edges.length} edge ${edges.length === 1 ? "type" : "types"}`}
+          issues={edges}
+        />
+      ) : null}
+    </AlertDialogBody>
+  );
+}
+
+function GeneralIssuesGroup({ issues }: { issues: GeneralImportIssue[] }) {
+  return (
+    <Group size="small">
+      <GroupHeader>
+        <GroupTitle>File structure</GroupTitle>
+      </GroupHeader>
+      {issues.map((issue, i) => (
+        <GroupItem key={i} className="gx-wrap-break-word space-y-1 text-xs">
+          <div>
+            <span className="font-mono">{issue.location}</span>: {issue.message}
+          </div>
+        </GroupItem>
+      ))}
+    </Group>
+  );
+}
+
+function EntryIssuesGroup({
+  label,
+  issues,
+}: {
+  label: string;
+  issues: EntryImportIssue[];
+}) {
+  return (
+    <Group size="small">
+      <GroupHeader>
+        <GroupTitle>{label}</GroupTitle>
+      </GroupHeader>
+      {issues.map((issue, i) => (
+        <GroupItem key={i} className="gx-wrap-break-word space-y-1 text-xs">
+          <div>
+            <span className="text-foreground font-mono font-medium">
+              {issue.typeName}
+            </span>
+            {" → "}
+            <span className="font-mono">{issue.field}</span>: {issue.message}
+          </div>
+          <div className="text-muted-foreground">
+            value:{" "}
+            <span className="font-mono">{formatIssueValue(issue.value)}</span>
+          </div>
+        </GroupItem>
+      ))}
+    </Group>
+  );
+}
+
+function formatIssueValue(value: unknown): string {
+  if (typeof value === "string") {
+    return `"${value}"`;
+  }
+  if (value === undefined) {
+    return "undefined";
+  }
+  return JSON.stringify(value);
+}
