+++ packages/graph-explorer/src/modules/StyleImport/LoadStylesButton.tsx
@@ -3,17 +3,11 @@ import { AlertTriangleIcon, FolderOpenIcon } from "lucide-react";
 import { startTransition, useActionState } from "react";
 import { toast } from "sonner";
 
-import type {
-  EntryImportIssue,
-  GeneralImportIssue,
-  ImportIssue,
-  StylingParseResult,
-} from "@/core/styling";
+import type { ImportIssue, StylingParseResult } from "@/core/styling";
 
 import {
   AlertDialog,
   AlertDialogAction,
-  AlertDialogBody,
   AlertDialogContent,
   AlertDialogDescription,
   AlertDialogFooter,
@@ -23,10 +17,6 @@ import {
   Button,
   Dialog,
   FileButton,
-  Group,
-  GroupHeader,
-  GroupItem,
-  GroupTitle,
 } from "@/components";
 import {
   userEdgeStylesAtom,
@@ -41,6 +31,7 @@ import {
 
 import type { StyleImportItem, StyleImportPlan } from "./styleImportPlan";
 
+import { ImportIssuesList } from "./ImportIssuesList";
 import { StyleImportModal } from "./StyleImportModal";
 import { buildStyleImportPlan } from "./styleImportPlan";
 import { useApplyStyleImport } from "./useApplyStyleImport";
@@ -209,7 +200,7 @@ function LoadInvalidContent({ issues }: { issues: ImportIssue[] }) {
           values below and try again.
         </AlertDialogDescription>
       </AlertDialogHeader>
-      <IssuesList issues={issues} />
+      <ImportIssuesList issues={issues} />
       <AlertDialogFooter>
         <AlertDialogAction>Close</AlertDialogAction>
       </AlertDialogFooter>
@@ -238,86 +229,4 @@ function LoadEmptyContent({ allMatched }: { allMatched: boolean }) {
       </AlertDialogFooter>
     </AlertDialogContent>
   );
-}
-
-function IssuesList({ issues }: { issues: ImportIssue[] }) {
-  const general = issues.filter(issue => issue.scope === "general");
-  const entries = issues.filter(issue => issue.scope === "entry");
-  const vertices = entries.filter(issue => issue.entityType === "vertex");
-  const edges = entries.filter(issue => issue.entityType === "edge");
-  return (
-    <AlertDialogBody className="space-y-3">
-      {general.length > 0 ? <GeneralIssuesGroup issues={general} /> : null}
-      {vertices.length > 0 ? (
-        <EntryIssuesGroup
-          label={`${vertices.length} node ${vertices.length === 1 ? "type" : "types"}`}
-          issues={vertices}
-        />
-      ) : null}
-      {edges.length > 0 ? (
-        <EntryIssuesGroup
-          label={`${edges.length} edge ${edges.length === 1 ? "type" : "types"}`}
-          issues={edges}
-        />
-      ) : null}
-    </AlertDialogBody>
-  );
-}
-
-function GeneralIssuesGroup({ issues }: { issues: GeneralImportIssue[] }) {
-  return (
-    <Group size="small">
-      <GroupHeader>
-        <GroupTitle>File structure</GroupTitle>
-      </GroupHeader>
-      {issues.map((issue, i) => (
-        <GroupItem key={i} className="gx-wrap-break-word space-y-1 text-xs">
-          <div>
-            <span className="font-mono">{issue.location}</span>: {issue.message}
-          </div>
-        </GroupItem>
-      ))}
-    </Group>
-  );
-}
-
-function EntryIssuesGroup({
-  label,
-  issues,
-}: {
-  label: string;
-  issues: EntryImportIssue[];
-}) {
-  return (
-    <Group size="small">
-      <GroupHeader>
-        <GroupTitle>{label}</GroupTitle>
-      </GroupHeader>
-      {issues.map((issue, i) => (
-        <GroupItem key={i} className="gx-wrap-break-word space-y-1 text-xs">
-          <div>
-            <span className="text-foreground font-mono font-medium">
-              {issue.typeName}
-            </span>
-            {" → "}
-            <span className="font-mono">{issue.field}</span>: {issue.message}
-          </div>
-          <div className="text-muted-foreground">
-            value:{" "}
-            <span className="font-mono">{formatIssueValue(issue.value)}</span>
-          </div>
-        </GroupItem>
-      ))}
-    </Group>
-  );
-}
-
-function formatIssueValue(value: unknown): string {
-  if (typeof value === "string") {
-    return `"${value}"`;
-  }
-  if (value === undefined) {
-    return "undefined";
-  }
-  return JSON.stringify(value);
 }
